new aura

+++ src/server/game/battle/skill.go
@@ -65,7 +65,7 @@ func (self *BattleSkill) Cast(u *BattleUnit, time int32) {
 	self.start_time = time
 	self.update_time = time
 	self.onStart()
-	self.owner.AddCampaignDetail(CampaignEvent_Cast, self.proto.Id, self.proto.Level, 0, 0)
+	self.owner.AddCampaignDetail(CampaignEvent_Cast, int32(self.proto.Id), int32(self.proto.Level), 0, 0)
 	fmt.Println(u.Name(), "释放了技能:", self.proto.Id, self.proto.Level)
 }
 
@@ -183,5 +183,5 @@ func (self *BattleSkill) do_attack(target *BattleUnit) {
 	if ctx.damage.crit {
 		is_crit = 1
 	}
-	target.AddCampaignDetail(CampaignEvent_Hurt, ctx.damage.hurt, is_crit, 0, 0)
+	target.AddCampaignDetail(CampaignEvent_Hurt, int32(ctx.damage.hurt), int32(is_crit), 0, 0)
 }
